--- example/bench/parameters.hpp
@@ -1,114 +1,0 @@
-#pragma once
-
-#include <exception>
-#include <fstream>
-#include <iostream>
-#include <string>
-
-#include <sup/json_params.hpp>
-
-#include <arbor/common_types.hpp>
-
-using arb::time_type;
-
-struct bench_params {
-    struct cell_params {
-        double spike_freq_hz = 10;   // Frequency in hz that cell will generate (poisson) spikes.
-        double realtime_ratio = 0.1; // Integration speed relative to real time, e.g. 10 implies
-                                     // that a cell is integrated 10 times slower than real time.
-    };
-    struct network_params {
-        unsigned fan_in = 5000;      // Number of incoming connections on each cell.
-        double min_delay = 10;       // Used as the delay on all connections.
-    };
-    std::string name = "default";    // Name of the model.
-    unsigned num_cells = 1000;       // Number of cells in model.
-    time_type duration = 100;          // Simulation duration in ms.
-
-    cell_params cell;                // Cell parameters for all cells in model.
-    network_params network;          // Description of the network.
-
-    // Expected simulation performance properties based on model parameters.
-
-    // Time to finish simulation if only cell overheads are counted.
-    double expected_advance_time() const {
-        return cell.realtime_ratio * duration*1e-3 * num_cells;
-    }
-    // Total expected number of spikes generated by simulation.
-    unsigned expected_spikes() const {
-        return num_cells * duration*1e-3 * cell.spike_freq_hz;
-    }
-    // Expected number of spikes generated per min_delay/2 interval.
-    unsigned expected_spikes_per_interval() const {
-        return num_cells * network.min_delay*1e-3/2 * cell.spike_freq_hz;
-    }
-    // Expected number of post-synaptic events delivered over simulation.
-    unsigned expected_events() const {
-        return expected_spikes() * network.fan_in;
-    }
-    // Expected number of post-synaptic events delivered per min_delay/2 interval.
-    unsigned expected_events_per_interval() const {
-        return expected_spikes_per_interval() * network.fan_in;
-    }
-};
-
-std::ostream& operator<<(std::ostream& o, const bench_params& p) {
-    o << "benchmark parameters:\n"
-      << "  name:          " << p.name << "\n"
-      << "  num cells:     " << p.num_cells << "\n"
-      << "  duration:      " << p.duration << " ms\n"
-      << "  fan in:        " << p.network.fan_in << " connections/cell\n"
-      << "  min delay:     " << p.network.min_delay << " ms\n"
-      << "  spike freq:    " << p.cell.spike_freq_hz << " Hz\n"
-      << "  cell overhead: " << p.cell.realtime_ratio << " ms to advance 1 ms\n";
-    o << "expected:\n"
-      << "  cell advance: " << p.expected_advance_time() << " s\n"
-      << "  spikes:       " << p.expected_spikes() << "\n"
-      << "  events:       " << p.expected_events() << "\n"
-      << "  spikes:       " << p.expected_spikes_per_interval() << " per interval\n"
-      << "  events:       " << p.expected_events_per_interval()/p.num_cells << " per cell per interval";
-    return o;
-}
-
-bench_params read_options(int argc, char** argv) {
-    using sup::param_from_json;
-
-    bench_params params;
-    if (argc<2) {
-        std::cout << "Using default parameters.\n";
-        return params;
-    }
-    if (argc>2) {
-        throw std::runtime_error("More than command line one option not permitted.");
-    }
-
-    std::string fname = argv[1];
-    std::cout << "Loading parameters from file: " << fname << "\n";
-    std::ifstream f(fname);
-
-    if (!f.good()) {
-        throw std::runtime_error("Unable to open input parameter file: "+fname);
-    }
-
-    nlohmann::json json;
-    json << f;
-
-    param_from_json(params.name, "name", json);
-    param_from_json(params.num_cells, "num-cells", json);
-    param_from_json(params.duration, "duration", json);
-    param_from_json(params.network.min_delay, "min-delay", json);
-    param_from_json(params.network.fan_in, "fan-in", json);
-    param_from_json(params.cell.realtime_ratio, "realtime-ratio", json);
-    param_from_json(params.cell.spike_freq_hz, "spike-frequency", json);
-
-    for (auto it=json.begin(); it!=json.end(); ++it) {
-        std::cout << "  Warning: unused input parameter: \"" << it.key() << "\"\n";
-    }
-    std::cout << "\n";
-
-    return params;
-}
-
-bench_params read_options(int argc, char** argv);
-
-std::ostream& operator<<(std::ostream& o, const bench_params& p);
